+++ lib/feature/admission/placement/ui/widgets/nom_scan_dialog.dart
@@ -514,7 +514,7 @@ class ChangeQuantity extends StatefulWidget {
   const ChangeQuantity(
       {super.key, required this.qty, required this.nom, required this.count});
 
-  final double qty;
+  final int qty;
   final double count;
   final AdmissionNom nom;
