+++ lib/feature/return/return_client/models/return_nom_dto.dart
@@ -44,7 +44,7 @@ class ReturnNomDTO {
 @JsonSerializable(createPerFieldToJson: false)
 class BarcodeDTO {
   final String? barcode;
-  final double? ratio;
+  final int? ratio;
 
   BarcodeDTO({required this.barcode, required this.ratio});
