+++ lib/feature/return/return_repository/models/noms_model.dart
@@ -40,7 +40,7 @@ class ReturnNom {
 
 class Barcode {
   final String barcode;
-  final double ratio;
+  final int ratio;
 
   Barcode({required this.barcode, required this.ratio});
