+++ lib/models/barcode_dto_model.dart
@@ -5,7 +5,7 @@ part 'barcode_dto_model.g.dart';
 @JsonSerializable()
 class BarcodeDTO {
   final String? barcode;
-  final double? ratio;
+  final int? ratio;
 
   BarcodeDTO({required this.barcode, required this.ratio});
